Update patch set 1
Patch Set 1:

RaymanFX: The SLTE has Android 5.0.2. The firmware is form LL ...

However, it skips the default firmware cause it is looking for a flavour. This patch works fine on my device and probably it didn't work before at all. I guess if there is only one firmware file the best is to set it in the config file.

Patch-set: 1