Update patch set 3
Patch Set 3:

> > Patch Set 3:
> > 
> > (1 comment)

The point I was trying to make was to keep authenticity of libbt-vendor for devices that don't use this hack. I understand that these are just prototype functions ,which shouldn't impact anything, but guarding them properly(just in case), wont do any bad.
If someone wants to use this functions, with BTHW_FW_EXTENTED_CONFIGURATION set to FALSE, then he/she should really proceede following the normal behavior, and do it by sending opcode command. If not for the hack these prototypes wouldn't be there anyway.

Patch-set: 3